requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
python-dotenv==0.19.0
supabase==1.0.3
python-jose[cryptography]==3.3.0
requests==2.26.0
openai>=1.3.0
crewai>=0.1.0
langchain>=0.1.0
langchain-community>=0.0.10
n8n>=0.176.0
asyncio>=3.4.3
python-jose>=3.3.0
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.0
pandas==2.2.0
plotly==5.18.0
numpy==1.26.3
aiohttp==3.9.3
redis==5.0.1
psycopg2-binary==2.9.9
bcrypt==4.1.2
pyjwt==2.8.0
google-auth==2.27.0
duckduckgo-search==4.1.1
passlib[bcrypt]==1.7.4
prometheus-client==0.19.0
sentry-sdk==1.39.1
backblaze-b2==1.4.0
cryptography==42.0.2
email-validator==2.1.0.post1
python-multipart==0.0.6
# Genix Business Ecosystem dependencies
stripe==7.5.0
anthropic==0.8.0
google-generativeai==0.3.1
tenacity==8.2.3